requirements.txt
# Core dependencies
click>=8.0.0
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
tabulate>=0.9.0
colorama>=0.4.6

# Kubernetes and cloud
kubernetes>=28.0.0
boto3>=1.34.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.19.0

# Data processing and analytics
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# Machine Learning - Production Models
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Authentication and security
pyjwt>=2.8.0
cryptography>=41.0.0
ldap3>=2.9.0
pyotp>=2.8.0
bcrypt>=4.0.0

# HTTP and API
requests>=2.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# FastAPI dependencies
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0

# Additional API server dependencies
starlette>=0.27.0
httpx>=0.24.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0
psutil>=5.9.0

# Data Processing
xmltodict>=0.13.0

# Cloud Authentication and APIs
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0
google-cloud-billing>=1.11.0
google-cloud-container>=2.30.0
google-cloud-compute>=1.14.0
google-cloud-bigquery>=3.11.0
azure-identity>=1.13.0
azure-mgmt-costmanagement>=3.0.0
azure-mgmt-containerservice>=27.0.0
azure-mgmt-compute>=30.0.0

# Caching and storage
redis>=4.6.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Build and packaging
pyinstaller>=6.0.0 